Rename ReportedError.js file in ReportedErrors.js and rename functions to create reported errors
